Lift the 401 landing rule out of axios into a transport-neutral module
